% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{get_pipeline}
\alias{get_pipeline}
\title{A simple targets pipeline}
\usage{
get_pipeline(suff)
}
\arguments{
\item{suff}{The suffix to the target name}
}
\value{
A list of targets
}
\description{
A simple targets pipeline
}
