% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{plus1}
\alias{plus1}
\title{A function for testing
Adds 1 to \code{x}.}
\usage{
plus1(x)
}
\arguments{
\item{x}{The incoming parameter}
}
\value{
\code{x + 1}
}
\description{
A function for testing
Adds 1 to \code{x}.
}
